+++ packages/generator-odata-downloader/jest.config.mjs
@@ -22,7 +22,8 @@ config.snapshotFormat = {
 };
 config.moduleNameMapper = {
     ...config.moduleNameMapper,
-    '^@vscode-logging/logger$': '<rootDir>/test/__mocks__/vscode-logging-logger.mjs'
+    '^@vscode-logging/logger$': '<rootDir>/test/__mocks__/vscode-logging-logger.mjs',
+    '^@sap/ux-cds-compiler-facade$': '<rootDir>/test/__mocks__/@sap/ux-cds-compiler-facade.ts'
 };
 config.transformIgnorePatterns = [
     'node_modules/(?!(?:.*?/)?(@sap-ux|@sap-ux-private|@sap/ux-cds-compiler-facade|@vscode-logging|@sap-devx[+/]yeoman-ui-types)/)'